% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dMI.R
\name{add.utility}
\alias{add.utility}
\title{Add utility to data}
\usage{
add.utility(
  data,
  dimnames = c("MO", "SC", "UA", "PD", "AD"),
  utilityname,
  impnum
)
}
\arguments{
\item{data}{The data containing the EQ-5D-5L responses}

\item{dimnames}{A character vector of length 5 giving the names in the dataset of the columns corresponding to the responses to mobility, self-care, usual activities, pain/discomfort and anxiety/depression, respectively. Defaults to c("MO","SC","UA","PD","AD")}

\item{utilityname}{The name to be given to the utility corresponding to each health state. This should be distinct from the other column names in the dataset.}

\item{impnum}{The imputation number from mu.pred3125 to use when looking up the utilities}
}
\value{
the original dataset provided plus one extra column named utilityname giving the utility for each corresponding to the health state
}
\description{
Add utility to data
}
\examples{
# simulate a dataset
 EQdata.sim<-array(dim=c(500,5))
colnames(EQdata.sim)<- c("MO", "SC", "UA", "PD", "AD")
EQdata.sim <- as.data.frame(EQdata.sim)
for (i in 1:nrow(EQdata.sim)){  EQdata.sim[i,] <- floor(runif(5, min=1, max=5))}
EQdata.sim$age <- round(abs(sample(rnorm(500, mean=65, sd=25))))
head(EQdata.sim)
EQdata.sim <- add.utility(data=EQdata.sim,dimnames=c("MO","SC","UA","PD","AD"),
                                                  utilityname="utility",impnum=1)
head(EQdata.sim)
}
