% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{allstates}
\alias{allstates}
\title{Look-up table assigning each of the EQ-5D-5L health states a number}
\format{
A dataframe.
}
\usage{
data(allstates)
}
\description{
Look-up table assigning each of the EQ-5D-5L health states a number
}
\keyword{datasets}
