% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dMI.R
\name{merge.dim}
\alias{merge.dim}
\title{Create 5-digit code indicating health state & look up in allstates}
\usage{
\method{merge}{dim}(
  data,
  dimnames = c("MO", "SC", "UA", "PD", "AD"),
  statename,
  allstates = allstates,
  stateidname
)
}
\arguments{
\item{data}{The data containing the EQ-5D-5L responses}

\item{dimnames}{A character vector of lenth 5 giving the names in the dataset of the columns corresponding to the responses to mobility, self-care, usual activities, pain/discomfort and anxiety/depression, respectively. Defaults to c("MO","SC","UA","PD","AD")}

\item{statename}{The name to be given to the 5-digit state created by the function}

\item{stateidname}{The name to be given to the state id column created by the function}
}
\value{
the original dataset provided plus two extra columns: one named statename with value equal to the five-digit code capturing the response to the EQ-5D-5L questionnaire, and another named stateidname, capturing the state id (i.e., a number from 1 to 3125) of the respondent's health state
}
\description{
Create 5-digit code indicating health state & look up in allstates
}
\examples{
# simulate a dataset
 EQdata.sim<-array(dim=c(500,5))
colnames(EQdata.sim)<- c("MO", "SC", "UA", "PD", "AD")
EQdata.sim <- as.data.frame(EQdata.sim)
for (i in 1:nrow(EQdata.sim)){  EQdata.sim[i,] <- floor(runif(5, min=1, max=5))}
EQdata.sim$age <- round(abs(sample(rnorm(500, mean=65, sd=25))))
EQdata.sim$cancer.site <- as.factor(sample(c("breast", "prostate", "colorectal", "other"), size = 500, replace = T, prob = c(0.21, 0.19, 0.12, 0.48)))
EQdata.sim$sex <- ifelse(EQdata.sim$cancer.site=="breast","female","male")
head(EQdata.sim)
EQdata.sim <- merge.dim(data=EQdata.sim,dimnames=c("MO","SC","UA","PD","AD"),statename="state",allstates=allstates,stateidname)
head(EQdata.sim)
}
