% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dMI.R
\name{mergedim}
\alias{mergedim}
\title{Create 5-digit code indicating health state and look up in allstates}
\usage{
mergedim(
  data,
  dimnames = c("MO", "SC", "UA", "PD", "AD"),
  statename,
  stateidname
)
}
\arguments{
\item{data}{The data containing the EQ-5D-5L responses}

\item{dimnames}{A character vector of length 5 giving the names in the dataset of the columns corresponding to the responses to mobility, self-care, usual activities, pain/discomfort and anxiety/depression, respectively. Defaults to c("MO","SC","UA","PD","AD")}

\item{statename}{The name to be given to the 5-digit state created by the function. This should be distinct from the other column names in the dataset.}

\item{stateidname}{The name to be given to the state id column created by the function. This should be distinct from the other column names in the dataset.}
}
\value{
the original dataset provided plus two extra columns: one named statename with value equal to the five-digit code capturing the response to the EQ-5D-5L questionnaire, and another named stateidname, capturing the state id (i.e., a number from 1 to 3125) of the respondent's health state
}
\description{
Create 5-digit code indicating health state and look up in allstates
}
\examples{
# simulate a dataset
 EQdata.sim<-array(dim=c(500,5))
colnames(EQdata.sim)<- c("MO", "SC", "UA", "PD", "AD")
EQdata.sim <- as.data.frame(EQdata.sim)
for (i in 1:nrow(EQdata.sim)){  EQdata.sim[i,] <- floor(runif(5, min=1, max=5))}
EQdata.sim$age <- round(abs(sample(rnorm(500, mean=65, sd=25))))
head(EQdata.sim)
EQdata.sim <- mergedim(data=EQdata.sim,dimnames=c("MO","SC","UA","PD","AD"),
                                                 statename="state",stateidname="stateid")
head(EQdata.sim)
}
