% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{mu.pred3125}
\alias{mu.pred3125}
\title{Multiply imputed value set for the EQ-5D-5L in Canada}
\format{
A matrix.
}
\usage{
data(mu.pred3125)
}
\description{
Multiply imputed value set for the EQ-5D-5L in Canada
}
\keyword{datasets}
