% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dMI.R
\name{multimp.eq5d}
\alias{multimp.eq5d}
\title{Create multiply imputed utilities, apply a user-defined function, and combine the results using Rubin's Rules}
\usage{
multimp.eq5d(
  data,
  userfx,
  dimnames,
  statename,
  stateidname,
  utilityname,
  numimp,
  ...
)
}
\arguments{
\item{data}{The data containing the EQ-5D-5L responses}

\item{userfx}{The function to be applied to the imputed utilities}

\item{dimnames}{A character vector of lenth 5 giving the names in the dataset of the columns corresponding to the responses to mobility, self-care, usual activities, pain/discomfort and anxiety/depression, respectively. Defaults to c("MO","SC","UA","PD","AD")}

\item{statename}{The name to be given to the 5-digit state created by the function}

\item{stateidname}{The name to be given to the state id column created by the function}

\item{utilityname}{The name to be given to the utility corresponding to each health state}

\item{numimp}{The number of imputations to use}

\item{...}{Additional arguments to userfx}
}
\value{
The result of using multiple imputation to apply userfx to the data
}
\description{
Create multiply imputed utilities, apply a user-defined function, and combine the results using Rubin's Rules
}
\examples{
# simulate a dataset
 EQdata.sim<-array(dim=c(500,5))
colnames(EQdata.sim)<- c("MO", "SC", "UA", "PD", "AD")
EQdata.sim <- as.data.frame(EQdata.sim)
for (i in 1:nrow(EQdata.sim)){  EQdata.sim[i,] <- floor(runif(5, min=1, max=5))}
EQdata.sim$age <- round(abs(sample(rnorm(500, mean=65, sd=25))))
head(EQdata.sim)

mean.sd.fn <- function(data){
mean.sd <- c(mean(data$utility), sd(data$utility)/sqrt(nrow(data)))
return(list(estimate=mean.sd[1],se=mean.sd[2]))
}
multimp.eq5d(data=EQdata.sim,userfx=mean.sd.fn,dimnames=c("MO","SC","UA","PD","AD"),
            statename="state",stateidname="stateid",utilityname="utility",numimp=100)
}
