% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dMI.R
\name{singleimp}
\alias{singleimp}
\title{use a single imputation to impute utilities, and apply a user-defined function to the resulting dataset}
\usage{
singleimp(
  impnum,
  data,
  userfx,
  dimnames,
  statename,
  stateidname,
  utilityname,
  ...
)
}
\arguments{
\item{impnum}{The imputation number from mu.pred3125 to use when looking up the utilities}

\item{data}{The data containing the EQ-5D-5L responses}

\item{userfx}{The function to be applied to the imputed utilities}

\item{dimnames}{A character vector of length 5 giving the names in the dataset of the columns corresponding to the responses to mobility, self-care, usual activities, pain/discomfort and anxiety/depression, respectively. Defaults to c("MO","SC","UA","PD","AD")}

\item{statename}{The name to be given to the 5-digit state created by the function. This should be distinct from the other column names in the dataset.}

\item{stateidname}{The name to be given to the state id column created by the function. This should be distinct from the other column names in the dataset.}

\item{utilityname}{The name to be given to the utility corresponding to each health state. This should be distinct from the other column names in the dataset.}

\item{...}{Additional arguments to userfx}
}
\value{
The result of applying userfx to the dataset using a single imputed value set
}
\description{
use a single imputation to impute utilities, and apply a user-defined function to the resulting dataset
}
\examples{
# simulate a dataset
 EQdata.sim<-array(dim=c(500,5))
colnames(EQdata.sim)<- c("MO", "SC", "UA", "PD", "AD")
EQdata.sim <- as.data.frame(EQdata.sim)
for (i in 1:nrow(EQdata.sim)){  EQdata.sim[i,] <- floor(runif(5, min=1, max=5))}
EQdata.sim$age <- round(abs(sample(rnorm(500, mean=65, sd=25))))
head(EQdata.sim)

mean.sd.fn <- function(data){
mean.sd <- c(mean(data$utility), sd(data$utility)/sqrt(nrow(data)))
return(list(estimate=mean.sd[1],se=mean.sd[2]))
}
singleimp(impnum=1,data=EQdata.sim,userfx=mean.sd.fn,dimnames=c("MO","SC","UA","PD","AD"),
                             statename="state",stateidname="stateid",utilityname="utility")
}
